[CHORE] Remove calls to jsonbin (#4200)
In the beginning of Hedy, we logged to jsonbin for simplicity, but we have since professionalized and now also log to s3. We can stop to also log to jsonbin (we did both because I was scared of losing logs and then we never removed jsonbin)
